chore: renamed BaseVoucherView to VoucherCardView and moved DokuVoucherView.Model declaration to a separate file
